LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/shape-outside/values/shape-outside-circle-005-expected.txt


PASS test unit: cm - circle(50cm at 50% 50%) - inline
FAIL test unit: cm - circle(50cm) - inline assert_equals: expected "circle(50cm)" but got "circle(50cm at 50% 50%)"
PASS test unit: mm - circle(50mm at 50% 50%) - inline
FAIL test unit: mm - circle(50mm) - inline assert_equals: expected "circle(50mm)" but got "circle(50mm at 50% 50%)"
PASS test unit: in - circle(50in at 50% 50%) - inline
FAIL test unit: in - circle(50in) - inline assert_equals: expected "circle(50in)" but got "circle(50in at 50% 50%)"
PASS test unit: pt - circle(50pt at 50% 50%) - inline
FAIL test unit: pt - circle(50pt) - inline assert_equals: expected "circle(50pt)" but got "circle(50pt at 50% 50%)"
PASS test unit: pc - circle(50pc at 50% 50%) - inline
FAIL test unit: pc - circle(50pc) - inline assert_equals: expected "circle(50pc)" but got "circle(50pc at 50% 50%)"
PASS test unit: em - circle(50em at 50% 50%) - inline
FAIL test unit: em - circle(50em) - inline assert_equals: expected "circle(50em)" but got "circle(50em at 50% 50%)"
PASS test unit: ex - circle(50ex at 50% 50%) - inline
FAIL test unit: ex - circle(50ex) - inline assert_equals: expected "circle(50ex)" but got "circle(50ex at 50% 50%)"
PASS test unit: ch - circle(50ch at 50% 50%) - inline
FAIL test unit: ch - circle(50ch) - inline assert_equals: expected "circle(50ch)" but got "circle(50ch at 50% 50%)"
PASS test unit: rem - circle(50rem at 50% 50%) - inline
FAIL test unit: rem - circle(50rem) - inline assert_equals: expected "circle(50rem)" but got "circle(50rem at 50% 50%)"
PASS test unit: vw - circle(50vw at 50% 50%) - inline
FAIL test unit: vw - circle(50vw) - inline assert_equals: expected "circle(50vw)" but got "circle(50vw at 50% 50%)"
PASS test unit: vh - circle(50vh at 50% 50%) - inline
FAIL test unit: vh - circle(50vh) - inline assert_equals: expected "circle(50vh)" but got "circle(50vh at 50% 50%)"
PASS test unit: vmin - circle(50vmin at 50% 50%) - inline
FAIL test unit: vmin - circle(50vmin) - inline assert_equals: expected "circle(50vmin)" but got "circle(50vmin at 50% 50%)"
PASS test unit: vmax - circle(50vmax at 50% 50%) - inline
FAIL test unit: vmax - circle(50vmax) - inline assert_equals: expected "circle(50vmax)" but got "circle(50vmax at 50% 50%)"
PASS test unit: cm - circle(50cm at 50% 50%) - computed
FAIL test unit: cm - circle(50cm) - computed assert_equals: expected "circle(1889.76px)" but got "circle(1889.76px at 50% 50%)"
PASS test unit: mm - circle(50mm at 50% 50%) - computed
FAIL test unit: mm - circle(50mm) - computed assert_equals: expected "circle(188.976px)" but got "circle(188.976px at 50% 50%)"
PASS test unit: in - circle(50in at 50% 50%) - computed
FAIL test unit: in - circle(50in) - computed assert_equals: expected "circle(4800px)" but got "circle(4800px at 50% 50%)"
PASS test unit: pt - circle(50pt at 50% 50%) - computed
FAIL test unit: pt - circle(50pt) - computed assert_equals: expected "circle(66.6667px)" but got "circle(66.6667px at 50% 50%)"
PASS test unit: pc - circle(50pc at 50% 50%) - computed
FAIL test unit: pc - circle(50pc) - computed assert_equals: expected "circle(800px)" but got "circle(800px at 50% 50%)"
PASS test unit: em - circle(50em at 50% 50%) - computed
FAIL test unit: em - circle(50em) - computed assert_equals: expected "circle(800px)" but got "circle(800px at 50% 50%)"
PASS test unit: ex - circle(50ex at 50% 50%) - computed
FAIL test unit: ex - circle(50ex) - computed assert_equals: expected "circle(640px)" but got "circle(640px at 50% 50%)"
PASS test unit: ch - circle(50ch at 50% 50%) - computed
FAIL test unit: ch - circle(50ch) - computed assert_equals: expected "circle(800px)" but got "circle(800px at 50% 50%)"
PASS test unit: rem - circle(50rem at 50% 50%) - computed
FAIL test unit: rem - circle(50rem) - computed assert_equals: expected "circle(800px)" but got "circle(800px at 50% 50%)"
PASS test unit: vw - circle(50vw at 50% 50%) - computed
FAIL test unit: vw - circle(50vw) - computed assert_equals: expected "circle(400px)" but got "circle(400px at 50% 50%)"
PASS test unit: vh - circle(50vh at 50% 50%) - computed
FAIL test unit: vh - circle(50vh) - computed assert_equals: expected "circle(300px)" but got "circle(300px at 50% 50%)"
PASS test unit: vmin - circle(50vmin at 50% 50%) - computed
FAIL test unit: vmin - circle(50vmin) - computed assert_equals: expected "circle(300px)" but got "circle(300px at 50% 50%)"
PASS test unit: vmax - circle(50vmax at 50% 50%) - computed
FAIL test unit: vmax - circle(50vmax) - computed assert_equals: expected "circle(400px)" but got "circle(400px at 50% 50%)"
